refactor: barrel exports for components
