+++ src/components/index.js
@@ -1,0 +1,4 @@
+export * from "./header";
+export * from "./nav";
+export * from "./footer";
+export * from "./submit-button";
